WIP: node-geocoder integration
trying to retrieve addresses, geocode them, and display a marker on the map.  Currently running into a bug with node-geocoder and the version of bluebird it is on resulting in an isIPV4 exception. Might be better to call google geocoder directly.  Pushing up this code for others to look at if they would like to.

Also changed html title to Denver Re:Imagine