Update defaultZoom and defaultCenter in MapCard
This pull request is to address issue #465. The goal is to ensure that any map we show on the site will have enough of a zoom to allow users to see the area they searched and that the map shown is centered on the location that the user searched. I left the default values to center on Denver in place so that we will always have a reasonable fallback.